+++ test/multiapi/Expected/AcceptanceTests/MultiapiWithSubmodule/multiapiwithsubmodule/submodule/_operations_mixin.py
@@ -17,7 +17,8 @@ from azure.core.paging import ItemPaged
 from azure.core.pipeline import PipelineResponse
 from azure.core.pipeline.transport import HttpRequest, HttpResponse
 from azure.core.polling import LROPoller, NoPolling, PollingMethod
-from azure.core.polling.base_polling import LROBasePolling
+from azure.mgmt.core.exceptions import ARMErrorFormat
+from azure.mgmt.core.polling.arm_polling import ARMPolling
 
 if TYPE_CHECKING:
     # pylint: disable=unused-import,ungrouped-imports
@@ -56,6 +57,40 @@ class MultiapiServiceClientOperationsMixin(object):
         mixin_instance._serialize = Serializer(self._models_dict(api_version))
         mixin_instance._deserialize = Deserializer(self._models_dict(api_version))
         return mixin_instance.begin_test_lro(product, **kwargs)
+
+    def begin_test_lro_and_paging(
+        self,
+        client_request_id=None,  # type: Optional[str]
+        test_lro_and_paging_options=None,  # type: Optional["models.TestLroAndPagingOptions"]
+        **kwargs  # type: Any
+    ):
+        """A long-running paging operation that includes a nextLink that has 10 pages.
+
+        :param client_request_id:
+        :type client_request_id: str
+        :param test_lro_and_paging_options: Parameter group.
+        :type test_lro_and_paging_options: ~multiapiwithsubmodule.submodule.v1.models.TestLroAndPagingOptions
+        :keyword callable cls: A custom type or function that will be passed the direct response
+        :keyword str continuation_token: A continuation token to restart a poller from a saved state.
+        :keyword polling: True for ARMPolling, False for no polling, or a
+         polling object for personal polling strategy
+        :paramtype polling: bool or ~azure.core.polling.PollingMethod
+        :keyword int polling_interval: Default waiting time between two polls for LRO operations if no Retry-After header is present.
+        :return: An instance of LROPoller that returns an iterator like instance of either PagingResult or the result of cls(response)
+        :rtype: ~azure.core.polling.LROPoller[~azure.core.paging.ItemPaged[~multiapiwithsubmodule.submodule.v1.models.PagingResult]]
+        :raises ~azure.core.exceptions.HttpResponseError:
+        """
+        api_version = self._get_api_version('begin_test_lro_and_paging')
+        if api_version == '1.0.0':
+            from .v1.operations import MultiapiServiceClientOperationsMixin as OperationClass
+        else:
+            raise NotImplementedError("APIVersion {} is not available".format(api_version))
+        mixin_instance = OperationClass()
+        mixin_instance._client = self._client
+        mixin_instance._config = self._config
+        mixin_instance._serialize = Serializer(self._models_dict(api_version))
+        mixin_instance._deserialize = Deserializer(self._models_dict(api_version))
+        return mixin_instance.begin_test_lro_and_paging(client_request_id, test_lro_and_paging_options, **kwargs)
 
     def test_one(
         self,