+++ test/vanilla/AcceptanceTests/test_additional_properties.py
@@ -147,3 +147,13 @@ class TestAdditionalProperties(object):
         output_ap_str_add = client.pets.create_ap_in_properties_with_ap_string(input_ap_str_add)
         assert output_ap_str_add.additional_properties['color'] ==  'red'
         assert output_ap_str_add.additional_properties1['weight'] ==  599
+
+    def test_models(self):
+        from additionalproperties.models import Error
+
+        if sys.version_info >= (3,5):
+            from additionalproperties.models._models_py3 import Error as ErrorPy3
+            assert Error == ErrorPy3
+        else:
+            from additionalproperties.models._models import Error as ErrorPy2
+            assert Error == ErrorPy2