+++ test/vanilla/AcceptanceTests/test_extensible_enums.py
@@ -72,3 +72,13 @@ class TestExtensibleEnums(object):
         assert returned_pet.days_of_week ==  "Friday"
         assert returned_pet.int_enum ==  "3"
         assert returned_pet.name ==  "Retriever"
+
+    def test_models(self):
+        from extensibleenumsswagger.models import Pet
+
+        if sys.version_info >= (3,5):
+            from extensibleenumsswagger.models._models_py3 import Pet as PetPy3
+            assert Pet == PetPy3
+        else:
+            from extensibleenumsswagger.models._models import Pet as PetPy2
+            assert Pet == PetPy2